pyproject.toml
[project]
name = "short-film"
version = "0.1.0"
description = "CLI tool for generating 60-second short films via AI video APIs"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.9"
authors = [
    { name = "Aaron Levin", email = "awlevin@comcast.net" },
]
keywords = ["ai", "video", "film", "cli", "openai", "sora", "gemini"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "Intended Audience :: End Users/Desktop",
    "Topic :: Multimedia :: Video",
    "Topic :: Artistic Software",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "typer>=0.9.0",
    "rich>=13.0.0",
    "openai>=1.0.0",
    "google-generativeai>=0.3.0",
    "pillow>=10.0.0",
    "pydantic>=2.0.0",
    "requests>=2.31.0",
    "tenacity>=8.2.0",
    "opencv-python>=4.8.0",
]

[project.scripts]
short-film = "short_film.cli:app"

[project.urls]
Homepage = "https://github.com/Olafs-World/short-film"
Repository = "https://github.com/Olafs-World/short-film"
Issues = "https://github.com/Olafs-World/short-film/issues"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["short_film"]

[tool.ruff]
line-length = 100
select = ["E", "F", "I", "N", "W"]
ignore = ["E501"]

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
markers = [
    "integration: marks tests as integration tests (deselect with '-m \"not integration\"')",
]

[dependency-groups]
dev = [
    "ruff>=0.1.0",
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.12.0",
]
